+++ src/revlanguage/workspace/VariableSlot.h
@@ -54,7 +54,7 @@ public:
 private:
     // helper methods
     RbPtr<Variable>                         getVectorizedVariable(const std::vector<int> &indices) const;
-    void                                    resetNames(const std::string &l, int level, int &index);
+    void                                    resetNames(const std::string &l, size_t level, int &index);
     void                                    resetVariables(std::vector<RbPtr<Variable> > &v, std::vector<int> &l, std::vector<int> &i);
